% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel.R
\name{ggfunnelESH}
\alias{ggfunnelESH}
\title{Funnel Plot}
\usage{
ggfunnelESH(model, xlabel, ylabel)
}
\arguments{
\item{model}{metafor rma model}

\item{xlabel}{x-axis label}

\item{ylabel}{y-axis label}
}
\description{
Generate a funnel plot
}
